api/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["skip"]

[tool.poetry.dependencies]
python = "^3.7"
mysqlclient = "^2.1.0"
SQLAlchemy = "^1.4.32"
uvicorn = "^0.17.6"
fastapi = "^0.75.0"
PyMySQL = "^1.0.2"
alembic = "^1.7.7"
graphene = "2.1.9"
graphene-sqlalchemy = "^2.3.0"
python-multipart = "^0.0.5"
passlib = "^1.7.4"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"